UPSTREAM: PCI: cadence: Move PCIe RP common functions to a separate file
Move the Cadence PCIe controller RP common functions into a separate file.
The common library functions are split from legacy PCIe RP controller
functions to a separate file.

[mani: removed the unused variable]
Link: https://patch.msgid.link/20251108140305.1120117-4-hans.zhang@cixtech.com
(cherry picked from commit b80a7b4713c967479752ea4801eb1d1933093f58)
